+++ CHANGELOG.md
@@ -10,13 +10,17 @@ Extending the adopted spec, each change should have a link to its corresponding 
 
 ## [Unreleased]
 
-### Added
+### Changed
 - Support for Terraform 0.12
+
+### Added
+- Add outputs `instance_name`, `instance_zone`
 - Add Kitchen-terraform tests
 - Update examples to fit terraform-google-log-export reformat
 
 ### Removed
 - Removed bats tests (obsolete)
+- Removed output `instance_id`
 
 ## [0.1.0] - 2019-03-29