+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Extending the adopted spec, each change should have a link to its corresponding 
 - Support for Terraform 0.12
 
 ### Added
+- Convert string variable `applications` to a list of strings
 - Add outputs `instance_name`, `instance_zone`
 - Add Kitchen-terraform tests
 - Update examples to fit terraform-google-log-export reformat